requirements.txt
absl-py==0.13.0
ahk==0.13.0
ahk-binary==1.1.33.9
ailment==9.0.6642
aiohttp==3.7.4.post0
alabaster==0.7.8
altgraph==0.17.2
angr==9.0.6642
archinfo==9.0.6642
asgiref==3.4.0
astunparse==1.6.3
async-api-client==0.0.4
async-timeout==3.0.1
atomicwrites==1.1.5
attrs==19.3.0
Automat==0.8.0
Babel==2.6.0
backcall==0.1.0
backports.entry-points-selectable==1.1.1
bcrypt==3.2.0
beautifulsoup4==4.10.0
binwalk==2.2.0
bitstring==3.1.7
bleach==3.1.1
blinker==1.4
brial==1.0.2
cachetools==4.2.2
capstone==4.0.2
certifi==2021.5.30
cffi==1.14.3
chardet==3.0.4
charset-normalizer==2.0.12
claripy==9.0.6642
cle==9.0.6642
Click==7.0
cloud-init==21.4
cloudpickle==1.6.0
cloudscraper==1.2.48
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
CppHeaderParser==2.7.4
cryptography==2.8
cvxopt==1.2.3
cycler==0.10.0
cypari2==2.1.1
Cython==0.29.14
DateTime==4.3
dbus-python==1.2.16
decorator==4.4.2
defusedxml==0.6.0
dirsearch==0.4.1
discord-components==2.1.2
discord-ui==5.1.6
discord.py==1.7.2
distlib==0.3.3
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.2.4
docutils==0.16
dpkt==1.9.5
entrypoints==0.3
expiringdict==1.2.1
filelock==3.4.0
flatbuffers==1.12
fpylll==0.5.1.dev0
future==0.18.2
gast==0.4.0
gitdb==4.0.7
GitPython==3.1.14
gmpy2==2.1.0b3
google-auth==1.32.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
greenlet==0.4.17
grpcio==1.34.1
gym==0.18.3
h5py==3.1.0
hentai==3.2.9
html5lib==1.0.1
httplib2==0.14.0
hyperlink==19.0.0
idna==2.10
image==1.5.33
image-slicer==2.1.1
imagesize==1.2.0
importlib-metadata==1.5.0
incremental==16.10.1
intervaltree==3.1.0
ipykernel==5.5.0
ipython==7.13.0
ipython_genutils==0.2.0
ipywidgets==6.0.0
itanium-demangler==1.0
itsdangerous==1.1.0
jedi==0.15.2
jikuai==0.0.8
Jinja2==2.10.1
joblib==1.0.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
jupyter-client==6.1.2
jupyter-core==4.6.3
Keras==2.4.3
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libnum==1.7.1
Mako==1.1.3
Markdown==3.3.4
MarkupSafe==1.1.0
matplotlib==3.1.2
mistune==0.8.4
more-itertools==4.2.0
MouseInfo==0.1.3
mpmath==1.1.0
msgpack==1.0.0
mulpyplexer==0.9
multidict==5.1.0
nbconvert==5.6.1
nbformat==5.0.4
netifaces==0.10.4
networkx==2.4
nose==1.3.7
notebook==6.0.3
numpy==1.19.5
oauthlib==3.2.0
olefile==0.46
opencv-python==4.5.2.54
opt-einsum==3.3.0
packaging==20.4
pandas==1.2.3
pandocfilters==1.4.2
paramiko==2.7.2
parso==0.5.2
path.py==12.0.1
pefile==2019.4.18
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.2.0
pixiv-api==1.0.0
PixivPy==3.5.11
pkgconfig==1.5.1
platformdirs==2.4.0
pluggy==0.13.0
plumbum==1.7.0
ply==3.11
pplpy==0.8.4
primefac==1.1
progressbar2==3.53.1
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.17.3
psutil==5.7.2
-e git+https://github.com/Gallopsled/pwntools@72ca730f4b036b1f01cdf5dffcbebf0625f67426#egg=pwntools
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycparser==2.20
pycrypto==2.6.1
pydot==1.4.2
pyelftools==0.26
PyGetWindow==0.0.9
pyglet==1.5.14
Pygments==2.7.1
PyGObject==3.36.0
PyHamcrest==1.9.0
pyinotify==0.9.6
pyinstaller==4.7
pyinstaller-hooks-contrib==2021.3
PyJWT==1.7.1
pymacaroons==0.13.0
PyMsgBox==1.0.9
PyNaCl==1.3.0
pynvim==0.4.2
PyOpenGL==3.1.0
pyOpenSSL==19.0.0
pyparsing==2.4.7
pyperclip==1.8.2
PyQt5==5.14.1
pyqtgraph==0.11.0rc0
PyRect==0.1.4
pyrsistent==0.15.5
PyScreeze==0.1.28
pyserial==3.4
PySMT==0.9.0
PySocks==1.7.1
PyStemmer==1.3.0
pytest==4.6.9
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
python-dotenv==0.17.1
python-utils==2.5.6
python3-openid==3.1.0
python3-xlib==0.15
pytweening==1.0.4
pytz==2019.3
pyvex==9.0.6642
PyYAML==5.3.1
pyzmq==18.1.1
raw-zlib==0.1.15
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
roman==2.0.0
ROPGadget==6.3
rpy2==3.2.6
rpyc==5.0.1
rsa==4.7.2
sage==9.0
sagenb-export==3.2
saucenao-api==2.4.0
scikit-learn==0.24.1
scipy==1.3.3
seaborn==0.11.1
SecretStorage==2.3.1
Send2Trash==1.5.0
service-identity==18.1.0
setuptools-scm==3.4.3
simplegeneric==0.8.1
simplejson==3.16.0
sip==4.19.21
six==1.15.0
smmap==4.0.0
snowballstemmer==2.0.0
sortedcontainers==2.2.2
sos==4.2
soupsieve==2.3.1
Sphinx==1.8.5
sqlparse==0.4.1
ssh-import-id==5.10
sympy==1.5.1
systemd-python==234
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.5.0
tensorflow-gpu==2.5.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.0
tornado==5.1.1
tqdm==4.62.3
traitlets==4.3.3
tweepy==4.6.0
Twisted==18.9.0
typing-extensions==3.7.4.3
tzlocal==2.1b1
ubuntu-advantage-tools==27.6
ufw==0.36
unattended-upgrades==0.1
unicorn==1.0.2rc4
urllib3==1.25.11
vcversioner==2.16.0.0
virtualenv==20.10.0
virtualenvwrapper-win==1.2.6
wadllib==1.3.3
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==2.0.1
wget==3.2
widgetsnbextension==2.0.0
wrapt==1.12.1
yarl==1.6.3
youtube-dl==2021.12.17
z3-solver==4.8.10.0
zipp==1.0.0
zope.interface==4.7.1
